+++ src/main/groovy/com/bugsnag/android/gradle/BugsnagPluginExtension.groovy
@@ -6,8 +6,9 @@ package com.bugsnag.android.gradle
 class BugsnagPluginExtension {
     def String endpoint = 'https://upload.bugsnag.com'
     def String apiKey = null
-    def boolean autoUpload = true;
-    def boolean autoProguardConfig = true;
+    def boolean autoUpload = true
+    def boolean autoProguardConfig = true
+    def boolean uploadDebugBuildMappings = false
     def boolean overwrite = false
     def int retryCount = 0
     def boolean ndk = false